disable `trigger_file` option from recovery.conf
Triggering of a slave to stop replication and enable writing works.
However if a second slave (with data persistence) is reconfigured to
use the first slave (triggered) as the master, the second slave cannot
replicate the data.

At the moment the only way to convert a slave to a master is to stop it
and start it as a master in which case the above issue is not seen.

To avoid users from trying to trigger a slave, we removed the config
altogether until the issue can be resolved.